requirements.txt
accelerate==0.34.2
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
appnope==0.1.4
asttokens==2.4.1
attrs==24.2.0
blinker==1.8.2
blobfile==3.0.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
comm==0.2.2
contourpy==1.3.0
cryptography==43.0.1
cycler==0.12.1
debugpy==1.8.5
decorator==5.1.1
diffusers==0.23.0
dill==0.3.8
diskcache==5.6.3
einops==0.8.0
executing==2.1.0
filelock==3.16.0
fire==0.6.0
Flask==3.0.3
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.9.0
ftfy==6.2.3
huggingface-hub==0.24.6
idna==3.8
importlib_metadata==8.4.0
invisible-watermark==0.2.0
ipykernel==6.29.5
ipython==8.27.0
ipywidgets==8.1.5
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
kandinsky2 @ git+https://github.com/ai-forever/Kandinsky-2.git@aeefc1ce3a989eefe7c99d6a02cce44318c4d210
kiwisolver==1.4.7
lightning-utilities==0.11.7
llama_cpp_python==0.2.90
lxml==5.3.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
nest-asyncio==1.6.0
networkx==3.3
nltk==3.9.1
numpy==2.1.1
omegaconf==2.3.0
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.2
parso==0.8.4
pathos==0.3.2
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.2.2
pox==0.3.4
ppft==1.7.6.8
prompt_toolkit==3.0.47
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pycryptodomex==3.20.0
PyExecJS==1.5.1
Pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-pptx==1.0.2
pytorch-lightning==2.4.0
pytz==2024.1
PyWavelets==1.7.0
PyYAML==6.0.2
pyzmq==26.2.0
regex==2024.7.24
requests==2.32.3
safetensors==0.4.5
sentencepiece==0.2.0
six==1.16.0
stack-data==0.6.3
sympy==1.13.2
termcolor==2.4.0
tokenizers==0.19.1
torch==2.4.1
torchmetrics==1.4.1
torchvision==0.19.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.44.2
translators==5.9.2
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wcwidth==0.2.13
Werkzeug==3.0.4
widgetsnbextension==4.0.13
XlsxWriter==3.2.0
yarl==1.10.0
zipp==3.20.1
